Add Miner Route
